chore : folder move hooks => utils
hook에는 어울리지 않는 파일이라고 판단됨!
